requirements.txt
annotated-types==0.7.0
arabic-reshaper==3.0.0
asgiref==3.9.1
asn1crypto==1.5.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
certifi==2025.8.3
cffi==1.17.1
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.3
cloudinary==1.44.1
constantly==23.10.4
cryptography==45.0.7
cssselect2==0.8.0
daphne==4.2.1
Django==5.2.5
django-allauth==65.11.2
django-cloudinary-storage==0.3.0
django-environ==0.12.0
gunicorn==23.0.0
html5lib==1.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
lxml==6.0.0
msgpack==1.1.1
mysqlclient==2.2.7
oscrypto==1.3.0
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.9
pydantic_core==2.33.2
pyHanko==0.29.1
pyhanko-certvalidator==0.27.0
PyJWT==2.10.1
pyOpenSSL==25.2.0
pypdf==6.0.0
python-bidi==0.6.6
python-decouple==3.8
PyYAML==6.0.2
redis==6.4.0
reportlab==4.4.3
requests==2.32.4
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
Twisted==25.5.0
txaio==25.6.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
webencodings==0.5.1
whitenoise==6.7.0
xhtml2pdf==0.2.17
zope.interface==8.0
google-generativeai==0.3.2
